# document/v1/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "API Documentation",
    "description": "This is the API documentation for the project."
  },
  "host": "localhost:3000",
  "basePath": "/api/v1",
  "schemes": ["http"],
  "paths": {
    "/lookup/fruits": {
      "get": {
        "summary": "Get Fruits",
        "description": "Get list of all fruits",
        "responses": {
          "200": {
            "description": "List of fruits",
            "schema": {
              "type": "object",
              "properties": {
                "code": { "type": "integer", "example": 200 },
                "message": { "type": "string", "example": "found" },
                "data": { "type": "array", "items": { "type": "object" } }
              }
            }
          },
          "404": { "description": "Not found" },
          "500": { "description": "Server error" }
        }
      }
    },
    "/lookup/categories": {
      "get": {
        "summary": "Get Categories",
        "description": "Get list of all categories",
        "responses": {
          "200": {
            "description": "List of categories",
            "schema": {
              "type": "object",
              "properties": {
                "code": { "type": "integer", "example": 200 },
                "message": { "type": "string", "example": "found" },
                "data": { "type": "array", "items": { "type": "object" } }
              }
            }
          },
          "404": { "description": "Not found" },
          "500": { "description": "Server error" }
        }
      }
    },
    "/lookup/ingredients": {
      "get": {
        "summary": "Get Ingredients",
        "description": "Get list of all ingredients",
        "responses": {
          "200": {
            "description": "List of ingredients",
            "schema": {
              "type": "object",
              "properties": {
                "code": { "type": "integer", "example": 200 },
                "message": { "type": "string", "example": "found" },
                "data": { "type": "array", "items": { "type": "object" } }
              }
            }
          },
          "404": { "description": "Not found" },
          "500": { "description": "Server error" }
        }
      }
    },
    "/restaurant": {
      "get": {
        "summary": "Get Restaurants",
        "description": "Get paginated list of restaurants for the user",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "type": "integer",
            "minimum": 1,
            "default": 1,
            "description": "Page number"
          },
          {
            "in": "query",
            "name": "limit",
            "type": "integer",
            "minimum": 1,
            "default": 5,
            "description": "Items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "List of restaurants",
            "schema": {
              "type": "object",
              "properties": {
                "code": { "type": "integer", "example": 200 },
                "message": { "type": "string", "example": "restaurants found" },
                "data": {
                  "type": "array",
                  "items": { "$ref": "#/definitions/Restaurant" }
                }
              }
            }
          },
          "401": { "description": "Unauthorized - User ID not found" },
          "500": { "description": "Server error" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "User Login",
        "description": "Authenticate a user and return a token.",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "format": "email"
                },
                "password": {
                  "type": "string",
                  "minLength": 6
                },
                "device": {
                  "$ref": "#/definitions/Device"
                }
              },
              "required": ["email", "password"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful login",
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/auth/signup": {
      "post": {
        "summary": "User Registration",
        "description": "Register a new user based on role (User/Chef/Delivery_Guy)",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "oneOf": [
                {
                  "type": "object",
                  "properties": {
                    "role": { "type": "string", "enum": ["User"] },
                    "user": { "$ref": "#/definitions/UserBase" },
                    "address": { "$ref": "#/definitions/Address" }
                  },
                  "required": ["role", "user", "address"]
                },
                {
                  "type": "object",
                  "properties": {
                    "role": { "type": "string", "enum": ["Chef"] },
                    "user": { "$ref": "#/definitions/UserBase" },
                    "restaurant": { "$ref": "#/definitions/Restaurant" },
                    "restaurant_address": { "$ref": "#/definitions/Address" }
                  },
                  "required": ["role", "user", "restaurant", "restaurant_address"]
                },
                {
                  "type": "object",
                  "properties": {
                    "role": { "type": "string", "enum": ["Delivery_Guy"] },
                    "user": { "$ref": "#/definitions/UserBase" }
                  },
                  "required": ["role", "user"]
                }
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User registered successfully"
          },
          "400": {
            "description": "Bad request - validation error"
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "User Logout",
        "description": "Logout a logged in user",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": true,
            "type": "string",
            "description": "Bearer token"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully logged out"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/auth/forgot-password": {
      "post": {
        "summary": "Forgot Password",
        "description": "Request password reset",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string"
                }
              },
              "required": ["email"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reset instructions sent"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
    "/auth/verify": {
      "post": {
        "summary": "Verify OTP",
        "description": "Verify OTP code",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "otp": {
                  "type": "string"
                }
              },
              "required": ["otp"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OTP verified successfully"
          },
          "400": {
            "description": "Invalid OTP"
          }
        }
      }
    },
    "/address": {
      "get": {
        "summary": "Get User Addresses",
        "description": "Get all addresses for the authenticated user",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": true,
            "type": "string",
            "description": "Bearer token"
          }
        ],
        "responses": {
          "200": {
            "description": "List of addresses",
            "schema": {
              "type": "object",
              "properties": {
                "code": { "type": "integer", "example": 200 },
                "message": { "type": "string", "example": "addresses found" },
                "addresses": {
                  "type": "object",
                  "properties": {
                    "addresses": {
                      "type": "array",
                      "items": { "$ref": "#/definitions/Address" }
                    },
                    "default_address": { "$ref": "#/definitions/Address" }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "404": { "description": "No addresses found" }
        }
      },
      "post": {
        "summary": "Add New Address",
        "description": "Add a new address for the authenticated user",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": true,
            "type": "string",
            "description": "Bearer token"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "allOf": [
                { "$ref": "#/definitions/Address" },
                {
                  "properties": {
                    "make_default": { "type": "boolean" }
                  }
                }
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Address created successfully",
            "schema": {
              "type": "object",
              "properties": {
                "code": { "type": "integer", "example": 200 },
                "message": { "type": "string", "example": "address created" },
                "address": { "$ref": "#/definitions/Address" }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      },
      "put": {
        "summary": "Update Address",
        "description": "Update an existing address",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": true,
            "type": "string",
            "description": "Bearer token"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "address_id": { "type": "string" },
                "address": { "type": "string" },
                "street": { "type": "string" },
                "state": { "type": "string" },
                "pin_code": { "type": "string" },
                "appartment": { "type": "string" },
                "label": { "type": "string", "enum": ["Home", "Work", "Other"] },
                "latitude": { "type": "number" },
                "longitude": { "type": "number" }
              },
              "required": ["address_id"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Address updated successfully",
            "schema": {
              "type": "object",
              "properties": {
                "code": { "type": "integer", "example": 200 },
                "message": { "type": "string", "example": "Updated" }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Address not found" }
        }
      },
      "delete": {
        "summary": "Delete Address",
        "description": "Delete an address for the authenticated user",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": true,
            "type": "string",
            "description": "Bearer token"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "address_id": { "type": "string" }
              },
              "required": ["address_id"]
            }
          }
        ],
        "responses": {
          "204": { "description": "Address deleted successfully" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/address/change-default": {
      "put": {
        "summary": "Set Default Address",
        "description": "Set an address as the default address for the user",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": true,
            "type": "string",
            "description": "Bearer token"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "address_id": { "type": "string" }
              },
              "required": ["address_id"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Default address changed successfully",
            "schema": {
              "type": "object",
              "properties": {
                "code": { "type": "integer", "example": 200 },
                "message": { "type": "string", "example": "default address changed" }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Address not found" }
        }
      }
    }
  },
  "definitions": {
    "Address": {
      "type": "object",
      "properties": {
        "address": { "type": "string", "nullable": true },
        "street": { "type": "string", "nullable": true },
        "state": { "type": "string", "nullable": true },
        "pin_code": { "type": "string", "nullable": true },
        "appartment": { "type": "string", "nullable": true },
        "label": { "type": "string", "enum": ["Home", "Work", "Other"], "default": "Home" },
        "latitude": { "type": "number", "nullable": true },
        "longitude": { "type": "number", "nullable": true }
      }
    },
    "PaginationParams": {
      "type": "object",
      "properties": {
        "page": {
          "type": "integer",
          "minimum": 1,
          "default": 1
        },
        "limit": {
          "type": "integer",
          "minimum": 1,
          "default": 5
        }
      }
    },
    "UserBase": {
      "type": "object",
      "properties": {
        "full_name": { "type": "string", "minLength": 1 },
        "role": { "type": "string", "enum": ["User", "Chef", "Delivery_Guy"] },
        "email": { "type": "string", "format": "email" },
        "password": { "type": "string", "minLength": 6 },
        "profile_image": { "type": "string", "format": "uri", "nullable": true },
        "phone": { "type": "string", "minLength": 1 }
      },
      "required": ["full_name", "role", "email", "password", "phone"]
    },
    "Restaurant": {
      "type": "object",
      "properties": {
        "name": { "type": "string", "minLength": 1 },
        "description": { "type": "string" },
        "phone": { "type": "string", "minLength": 1 },
        "email": { "type": "string", "format": "email" },
        "cover_image": { "type": "string", "format": "uri", "nullable": true },
        "logo": { "type": "string", "format": "uri", "nullable": true },
        "delivery_time": { "type": "integer", "minimum": 1 },
        "delivery_fees": { "type": "number", "minimum": 0 },
        "is_open": { "type": "boolean", "default": true },
        "opening_time": { "type": "string" },
        "closing_time": { "type": "string" }
      },
      "required": ["name", "phone", "email"]
    },
    "Device": {
      "type": "object",
      "properties": {
        "device_type": { "type": "string" },
        "device_name": { "type": "string" },
        "os_version": { "type": "string" },
        "model_name": { "type": "string" },
        "ip": { "type": "string" },
        "app_version": { "type": "string" },
        "platform": { "type": "string" }
      }
    }
  }
}